backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Environment and configuration
python-dotenv==1.0.0
pydantic[email]==2.5.0

# File handling
python-multipart==0.0.6
aiofiles==23.2.1

# Email support
emails==0.6.0

# Date and time handling
python-dateutil==2.8.2

# HTTP client
httpx==0.25.2

# Validation
email-validator==2.1.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Production
gunicorn==21.2.0

# Monitoring and logging
structlog==23.2.0

# CORS
fastapi-cors==0.0.6

# QR Code generation
qrcode[pil]==7.4.2
Pillow==10.1.0

# Excel support (for reports)
openpyxl==3.1.2

# PDF generation
reportlab==4.0.7

# Redis (for caching and sessions)
redis==5.0.1

# Celery (for background tasks)
celery==5.3.4

# Prometheus metrics
prometheus-client==0.19.0

# Health checks
healthcheck==1.3.0
